+++ packages/xo-web/src/common/intl/messages.js
@@ -83,6 +83,9 @@ const messages = {
   newServerPage: 'Server',
   newImport: 'Import',
   xosan: 'XOSAN',
+  backupDeprecatedMessage:
+    'Backup is deprecated, use Backup NG instead to create new backups.',
+  backupNgNewPage: 'New backup NG',
   backupOverviewPage: 'Overview',
   backupNewPage: 'New',
   backupRemotesPage: 'Remotes',
@@ -353,6 +356,7 @@ const messages = {
   jobUserNotFound: "This job's creator no longer exists",
   backupUserNotFound: "This backup's creator no longer exists",
   redirectToMatchingVms: 'Click here to see the matching VMs',
+  migrateToBackupNg: 'Migrate to backup NG',
   noMatchingVms: 'There are no matching VMs!',
   allMatchingVms: '{icon} See the matching VMs ({nMatchingVms, number})',
   backupOwner: 'Backup owner',